move Network to its own app
