+++ lemarche/siaes/admin.py
@@ -1,14 +1,6 @@
 from django.contrib import admin
 
-from lemarche.siaes.models import SiaeNetwork, Siae
-
-
-@admin.register(SiaeNetwork)
-class SiaeNetworkAdmin(admin.ModelAdmin):
-    list_display = ["id", "name", "brand", "created_at"]
-    search_fields = ["id", "name", "brand"]
-
-    readonly_fields = ["created_at", "updated_at"]
+from lemarche.siaes.models import Siae
 
 
 @admin.register(Siae)
