+++ lemarche/siaes/models.py
@@ -7,21 +7,6 @@ from lemarche.siaes.validators import validate_post_code, validate_siret, valida
 from lemarche.api.models import Sector
 
 
-class SiaeNetwork(models.Model):
-    name = models.CharField(verbose_name="Nom", max_length=255)
-    brand = models.CharField(verbose_name="Enseigne", max_length=255, blank=True)
-    website = models.URLField(verbose_name="Site web", blank=True)
-    created_at = models.DateTimeField("Date de création", default=timezone.now)
-    updated_at = models.DateTimeField("Date de modification", auto_now=True)
-
-    class Meta:
-        verbose_name = "Réseau"
-        verbose_name_plural = "Réseaux"
-
-    def __str__(self):
-        return self.name
-
-
 class Siae(models.Model):
     READONLY_FIELDS_FROM_C1 = [
         "name", "brand", "siret", "naf", "email", "phone", "kind", "nature",  # "website", "presta_type"
@@ -129,7 +114,7 @@ class Siae(models.Model):
 
     sectors = models.ManyToManyField(Sector)
     networks = models.ManyToManyField(
-        "SiaeNetwork",
+        "networks.Network",
         verbose_name='Réseaux',
         related_name='siaes',
         blank=True)
